requirements.txt
pygame
python-chess

# Hand gesture volume control dependencies
opencv-python>=4.8.0
mediapipe>=0.10.0
numpy>=1.24.0

# Platform-specific audio control (Windows only)
pycaw>=20230407; platform_system=="Windows"
comtypes>=1.1.0; platform_system=="Windows"
